requirements.txt
alembic==1.8.1
async-timeout==4.0.2
attrs==19.3.0
Automat==0.8.0
backcall==0.1.0
blinker==1.4
cachelib==0.9.0
certifi==2019.11.28
chardet==3.0.4
click==8.1.3
cloud-init==22.4.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
decorator==4.4.2
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
Flask==2.2.2
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-Principal==0.4.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
greenlet==1.1.3.post0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==5.0.0
importlib-resources==5.10.0
incremental==16.10.1
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.15.2
Jinja2==3.1.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
Mako==1.2.3
MarkupSafe==2.1.1
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
parso==0.5.2
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
psycopg2==2.9.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.1
python-debian===0.1.36ubuntu1
python-dotenv==0.21.0
PyYAML==5.3.1
redis==4.4.2
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.4
SQLAlchemy==1.4.42
ssh-import-id==5.10
systemd-python==234
traitlets==4.3.3
Twisted==18.9.0
ubuntu-advantage-tools==27.12
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
wcwidth==0.1.8
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.9.0
zope.interface==4.7.1
